+++ android/src/main/java/com/plugin/flutter/zsdk/SGDParams.java
@@ -31,6 +31,7 @@ public class SGDParams
     public static final String KEY_FIRMWARE = "appl.name";
     public static final String KEY_LINK_OS_VERSION = "appl.link_os_version";
     public static final String KEY_PRINTER_DPI = "head.resolution.in_dpi";
+    public static final String KEY_DEVICE_PRINT_HEAD_RESOLUTION = "device.printhead.resolution";
 
     /** Actions */
     public static final String KEY_MANUAL_CALIBRATION = "ezpl.manual_calibration";