+++ example/ios/Flutter/flutter_export_environment.sh
@@ -6,4 +6,6 @@ export "FLUTTER_TARGET=/Users/krrigan/Flutter/FlutterProjects/zsdk/example/lib/m
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
 export "FLUTTER_FRAMEWORK_DIR=/Users/krrigan/Flutter/SDK/bin/cache/artifacts/engine/ios"
+export "FLUTTER_BUILD_NAME=1.0.0"
+export "FLUTTER_BUILD_NUMBER=15"
 export "TRACK_WIDGET_CREATION=true"